Final Push
Fixed code so that the page is viewed correctly (used correct code 'flex-flow'). Changed the horizontal navigation to be a flex box with no wrapping but still be in a row, while having a set height. The image shave set sizes so that they fit correctly on the page, as if they were on an actual page. When the page is printed the navigation, footer, and article disappear. The sheet section now has a changed width and placement so that. Each image now has a size that makes it fit on the page better. Changed page 1 to link to the main page because it was renamed.